requirements.txt
streamlit
langchain
langchain-community
openai
faiss-cpu
python-dotenv
langchain-openai